Many various tweaks
All GPIO that I plan on making easily accessible is now routed to the
edge of the board.
I am still debating whether I want to keep the UART0_TX/UART0_RX and RST
lines available, since for the most part they probably wouldn't need to
be accessed after development, and during the physical button and USB
could be used.